Main menu / pause menu
I edited what Lily did for the main menu and then created a pause menu when esc is pressed. Some buttons do not work (ie About). Quit uses the debug log to demonstrate to us on unity that it is working, but won't actually quit the program unless we create a build. :)